// @types/logs.d.ts
export type ErrorCodes =
  | "00005"
  | "00006"
  | "00007"
  | "00008"
  | "00011"
  | "00012"
  | "00017"
  | "00024"
  | "00025"
  | "00026"
  | "00032"
  | "00036"
  | "00043"
  | "00044"
  | "00045"
  | "00054"
  | "00061"
  | "00084"
  | "00085"
  | "00037"
  | "00087"
  | "00088"
  | "00001"
  | "00002"
  | "00003"
  | "00004"
  | "00014"
  | "00015"
  | "00016"
  | "00018";

export type ErrorLevel =
  | "External Error"
  | "Server Error"
  | "Client Error"
  | "Uncaught Exception"
  | "Unhandled Rejection"
  | "Process-Env";